requirements.txt
litellm>=1.50.0
qdrant-client>=1.12.0
jsonschema>=4.23.0
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
rich>=13.9.0
Pillow>=10.4.0
